+++ packages/flutter_state_notifier/test/state_notifier_provider_test.dart
@@ -2,10 +2,10 @@ import 'package:flutter/src/widgets/basic.dart';
 import 'package:flutter/src/widgets/framework.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_state_notifier/flutter_state_notifier.dart';
-import 'package:mockito/mockito.dart';
-import 'package:state_notifier/state_notifier.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
+import 'package:state_notifier/state_notifier.dart';
 
 import 'common.dart';